lark: remove defaults for database configuration
Using `ENV.fetch` ensures that missing config eagerly fails, instead of using
undesirable defaults. For the rake tasks, this means we should decline to load
the database tasks if the `:sql` adapter is not in use. This way, an environment
still doesn't need to provide settings that won't be used. A guard is added to
achive this.

+++ lark/lib/tasks/db.rake
@@ -1,41 +1,43 @@
 # frozen_string_literal: true
 
-require 'sequel'
-require 'sequel/adapters/postgresql'
-require 'sequel/core'
-require 'sequel_pg'
-require_relative '../../config/database'
-require 'logger'
+if ENV.fetch('EVENT_ADAPTER', :sql).to_sym == :sql
+  require 'sequel'
+  require 'sequel/adapters/postgresql'
+  require 'sequel/core'
+  require 'sequel_pg'
+  require_relative '../../config/database'
+  require 'logger'
 
-namespace :db do
-  desc 'Create Database ...'
-  task :create do
-    puts "Database connection info: #{DATABASE.inspect}"
-    connection = Sequel.connect(DATABASE, logger: Logger.new($stderr))
-    begin
-      connection.execute "CREATE DATABASE #{DATABASE[:database]}"
-      puts "Database #{DATABASE[:database]} is created."
-    rescue Sequel::DatabaseError => e
-      puts "Error creating database #{DATABASE[:database]}: #{e}"
+  namespace :db do
+    desc 'Create Database ...'
+    task :create do
+      puts "Database connection info: #{DATABASE.inspect}"
+      connection = Sequel.connect(DATABASE, logger: Logger.new($stderr))
+      begin
+        connection.execute "CREATE DATABASE #{DATABASE[:database]}"
+        puts "Database #{DATABASE[:database]} is created."
+      rescue Sequel::DatabaseError => e
+        puts "Error creating database #{DATABASE[:database]}: #{e}"
+      end
     end
-  end
 
-  desc 'Run migrations ...'
-  task :migrate, [:version] do |t, args|
-    Sequel.extension :migration
-    version = args[:version].to_i if args[:version] && t
-    Sequel.connect(DATABASE, logger: Logger.new($stderr)) do |db|
-      Sequel::Migrator.run(db, 'db/migrations', target: version)
+    desc 'Run migrations ...'
+    task :migrate, [:version] do |t, args|
+      Sequel.extension :migration
+      version = args[:version].to_i if args[:version] && t
+      Sequel.connect(DATABASE, logger: Logger.new($stderr)) do |db|
+        Sequel::Migrator.run(db, 'db/migrations', target: version)
+      end
+    rescue PG::ConnectionBad => e
+      puts "Database connection failed on: #{DATABASE.inspect}"
+      raise e
     end
-  rescue PG::ConnectionBad => e
-    puts "Database connection failed on: #{DATABASE.inspect}"
-    raise e
-  end
 
-  desc 'Drop Database ...'
-  task :drop do
-    new_connection = Sequel.connect(DATABASE, logger: Logger.new($stderr))
-    new_connection.execute "DROP DATABASE IF EXISTS #{DATABASE[:database]}"
-    puts "#{DATABASE[:database]} is dropped"
+    desc 'Drop Database ...'
+    task :drop do
+      new_connection = Sequel.connect(DATABASE, logger: Logger.new($stderr))
+      new_connection.execute "DROP DATABASE IF EXISTS #{DATABASE[:database]}"
+      puts "#{DATABASE[:database]} is dropped"
+    end
   end
 end
